testing

+++ philo/philo.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   philo.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gforns-s <gforns-s@student.42.fr>          +#+  +:+       +#+        */
+/*   By: codespace <codespace@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/17 13:30:39 by gforns-s          #+#    #+#             */
-/*   Updated: 2024/04/25 03:32:10 by gforns-s         ###   ########.fr       */
+/*   Updated: 2024/04/25 02:10:15 by codespace        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,10 +40,10 @@ typedef struct s_philo
 	bool			full;
 	pthread_t		tid;
 	struct s_data	*data;
-	uint64_t		lst_meal;
-	uint64_t		t_death;
-	uint64_t		t_eat;
-	uint64_t		t_sleep;
+	unsigned long long		lst_meal;
+	unsigned long long		t_death;
+	unsigned long long		t_eat;
+	unsigned long long		t_sleep;
 	int				nb_meal;
 	pthread_mutex_t	lock;
 	pthread_mutex_t	*r_fork;
@@ -53,13 +53,13 @@ typedef struct s_philo
 typedef struct s_data
 {
 	int				nb_philo;
-	uint64_t		t_death;
-	uint64_t		t_eat;
-	uint64_t		t_sleep;
+	unsigned long long		t_death;
+	unsigned long long		t_eat;
+	unsigned long long		t_sleep;
 	int				nb_meal;
 	int				cnt_finish;
 	pthread_mutex_t	*fork;	
-	uint64_t		t0;
+	unsigned long long		t0;
 	pthread_mutex_t	print;
 	pthread_mutex_t	lock;
 	pthread_mutex_t	dead;
@@ -72,16 +72,16 @@ int			ft_atoi(const char *str);
 int			init_philos(t_data *data);
 int			create_philos(t_data *data);
 void		*routine(void *aux);
-uint64_t	get_time_ms(void);
-uint64_t	get_curr_time_clean(t_data *data);
-uint64_t	eat_time(t_philo *philo);
-uint64_t	sleep_time(t_philo *philo);
-uint64_t	death_time(t_philo *philo);
-uint64_t	lst_meal_time(t_philo *philo);
+unsigned long long	get_time_ms(void);
+unsigned long long	get_curr_time_clean(t_data *data);
+unsigned long long	eat_time(t_philo *philo);
+unsigned long long	sleep_time(t_philo *philo);
+unsigned long long	death_time(t_philo *philo);
+unsigned long long	lst_meal_time(t_philo *philo);
 int			ft_is_dead(t_data *data);
-void		ft_usleep(uint64_t time);
+void		ft_usleep(unsigned long long time);
 void		philo_round(t_philo *philo);
-void		ft_print_p(t_philo *philo, uint64_t t, int p, char *s);
+void		ft_print_p(t_philo *philo, unsigned long long t, int p, char *s);
 int			create_one_philo(t_data *data);
 int			ft_strcmp(const char *s1, const char *s2);
 void		increase_finished(t_data *data);
@@ -89,9 +89,9 @@ int			ft_monitor(t_data *data);
 int			check_philo_status(t_philo *philo);
 int			ft_finished(t_data *data);
 int			get_finished(t_data *data);
-uint64_t	get_curr_time_clean_u(t_data *data);
-uint64_t	get_lst_meal(t_philo *philo);
-uint64_t	get_death_time(t_data *data);
+unsigned long long	get_curr_time_clean_u(t_data *data);
+unsigned long long	get_lst_meal(t_philo *philo);
+unsigned long long	get_death_time(t_data *data);
 bool	get_death_status(t_philo *philo);
 void	set_lst_meal(t_data *data);
